+++ src/cloze.rs
@@ -188,6 +188,7 @@ pub fn expand(
             image: None,
             image_back: None,
             frontend: None,
+            max_stage: None,
         });
     }